NYM-1569: (tauri) ksni tray fix (#5647)
